backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-generativeai
ccxt
pandas
ta-lib
python-dotenv
sqlalchemy
aiosqlite
